DO NOT MERGE (yet) OpenFlow port values are uint32, correcting the size in the model for that
The do not merge is because this patch is not tested yet.

+++ southbound/southbound-impl/src/main/java/org/opendaylight/ovsdb/southbound/ovsdb/transact/TerminationPointCreateCommand.java
@@ -57,9 +57,9 @@ public class TerminationPointCreateCommand implements TransactCommand {
             Interface ovsInterface = TyperUtils.getTypedRowWrapper(transaction.getDatabaseSchema(), Interface.class);
             ovsInterface.setName(terminationPoint.getName());
             ovsInterface.setType(SouthboundMapper.createOvsdbInterfaceType(terminationPoint.getInterfaceType()));
-            Integer ofPort = terminationPoint.getOfport();
+            Long ofPort = terminationPoint.getOfport();
             if (ofPort != null) {
-                ovsInterface.setOpenFlowPort(Sets.newHashSet(ofPort.longValue()));
+                ovsInterface.setOpenFlowPort(Sets.newHashSet(ofPort));
             }
             Integer ofPortRequest = terminationPoint.getOfportRequest();
             if (ofPortRequest != null) {
